added progressview stuff, currently has hardcoded values for testing. need to figure out how to put the progress bar IN each cell"

+++ ParseStarterProject/SavingsItem.swift
@@ -22,7 +22,7 @@ class SavingsItem: NSObject {
     var total:Int
     
     // progress amount
-    var progress:Double
+    var progress:Float
     
     // initializes item
     init(text:String, total:Int) {
